webthoitrang/requirements.txt
Django==5.0
fashion-clip
torch 
torchvision 
faiss-cpu
djangorestframework
numpy
pandas
gdown
